requirements.txt
alembic==1.12.0
fastapi==0.103.2
psycopg2==2.9.9
python-dotenv==1.0.0
SQLAlchemy==2.0.21
uvicorn==0.23.2
